Create Recon-NgBulkAppending
Small script I'd written to automate the process of importing many domains/hosts at once to a Recon-ng database, instead of importing each of them one by one. 
There might be a more elegant, designated way of doing it (perhaps with sqlitebrowser) but this one works as well.
Naturally, change the parameters to specify your own file and database names).